pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "langgraph-helper-agent"
version = "0.1.0"
description = "AI assistant for LangGraph and LangChain documentation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core - LangGraph v1 + LangChain V1 API
    "langgraph>=1.0,<2",
    "langchain-core>=1.2,<2",
    "langchain-text-splitters>=1.1,<2",
    # LLM - Google Gemini
    "langchain-google-genai>=4,<5",
    # Vector store
    "chromadb>=1.4,<2",
    "langchain-chroma>=1.1,<2",
    # Web search
    "tavily-python>=0.5,<1",
    # Config & utilities
    "pydantic-settings>=2,<3",
    "tenacity>=9,<10",
    "typer>=0.21,<1",
    "rich>=14,<15",
    "httpx>=0.28,<1",
]

[project.scripts]
lha = "lha.main:main"

[project.optional-dependencies]
dev = [
    "mypy>=1.19",
    "ruff>=0.14",
    "python-dotenv>=1.0",
    "ipykernel>=6.29",
]

[tool.hatch.build.targets.wheel]
packages = ["src/lha"]

[tool.mypy]
python_version = "3.12"
strict = true

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.lint.per-file-ignores]
"notebooks/*.ipynb" = ["E402"]  # Allow imports anywhere in notebook cells

[tool.ruff.format]
quote-style = "double"
